+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+
+### Fixed
+- Reportable names doubling when loading from cache.
+
+
+
+
 ## [0.9.3] - 2023-03-21
 
 ### Fixed